Added proper fitness calculation

+++ Game/header/util.h
@@ -80,6 +80,6 @@ bool IsOnScreen(b2Vec2 position);
 float randomNumber(float start, float end);
 float lerp(float start, float end, float t);
 
-float fitness(int kills, double time_alive);
+float fitness(int kills, int time_alive);
 
 #endif
